+++ source/renderer/app/components/wallet/backup-recovery/MnemonicWord.js
@@ -1,9 +1,8 @@
 // @flow
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
-import classnames from 'classnames';
-import Button from 'react-polymorph/lib/components/Button';
-import SimpleButtonSkin from 'react-polymorph/lib/skins/simple/raw/ButtonSkin';
+import { Button } from 'react-polymorph/lib/components/Button';
+import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
 import styles from './MnemonicWord.scss';
 
 type Props = {
@@ -18,17 +17,16 @@ export default class MnemonicWord extends Component<Props> {
 
   render() {
     const { word, index, isActive, onClick } = this.props;
-    const componentClassNames = classnames([
-      'flat',
-      styles.component,
-      isActive ? styles.active : styles.inactive
-    ]);
+    const handleClick = onClick.bind(null, { word, index });
+
     return (
       <Button
-        className={componentClassNames}
+        className="flat"
+        themeOverrides={styles}
+        disabled={!isActive}
         label={word}
-        onClick={() => onClick({ word, index })}
-        skin={<SimpleButtonSkin />}
+        onClick={handleClick}
+        skin={ButtonSkin}
       />
     );
   }