+++ source/renderer/app/components/wallet/transactions/WalletTransactionsList.js
@@ -2,8 +2,8 @@
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
 import classnames from 'classnames';
-import Button from 'react-polymorph/lib/components/Button';
-import SimpleButtonSkin from 'react-polymorph/lib/skins/simple/raw/ButtonSkin';
+import { Button } from 'react-polymorph/lib/components/Button';
+import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
 import { defineMessages, intlShape } from 'react-intl';
 import moment from 'moment';
 import styles from './WalletTransactionsList.scss';
@@ -169,7 +169,7 @@ export default class WalletTransactionsList extends Component<Props> {
             className={buttonClasses}
             label={intl.formatMessage(messages.showMoreTransactionsButtonLabel)}
             onClick={this.onShowMoreTransactions.bind(this, walletId)}
-            skin={<SimpleButtonSkin />}
+            skin={ButtonSkin}
           />
         }
       </div>