+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 ## vNext
 =======
 
+### Chores
+
+- Integrates the latest features from the develop branch of react-polymorph. Features include: render props architecture, theme composition, and a ThemeProvider HOC. [PR 950](https://github.com/input-output-hk/daedalus/pull/950)
+
 ### Features
 
 - Implemented a switch instead of a link for “hide used” addresses on the Receive screen ([PR 935](https://github.com/input-output-hk/daedalus/pull/935))